+++ src/main/java/techcourse/myblog/application/CommentService.java
@@ -8,6 +8,9 @@ import techcourse.myblog.application.exception.NoArticleException;
 import techcourse.myblog.application.exception.NoUserException;
 import techcourse.myblog.application.exception.NotSameAuthorException;
 import techcourse.myblog.domain.*;
+import techcourse.myblog.domain.repository.ArticleRepository;
+import techcourse.myblog.domain.repository.CommentRepository;
+import techcourse.myblog.domain.repository.UserRepository;
 
 import javax.transaction.Transactional;
 import java.util.Collections;
