+++ src/main/java/techcourse/myblog/application/UserService.java
@@ -9,7 +9,7 @@ import techcourse.myblog.application.exception.EditException;
 import techcourse.myblog.application.exception.LoginException;
 import techcourse.myblog.application.exception.NoUserException;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.UserRepository;
+import techcourse.myblog.domain.repository.UserRepository;
 import techcourse.myblog.support.encrytor.EncryptHelper;
 
 import javax.transaction.Transactional;
