+++ src/main/java/techcourse/myblog/domain/repository/ArticleRepository.java
@@ -1,6 +1,7 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import techcourse.myblog.domain.Article;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
     boolean existsByTitle(String title);
