+++ src/main/java/techcourse/myblog/domain/repository/CommentRepository.java
@@ -1,6 +1,8 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import techcourse.myblog.domain.Article;
+import techcourse.myblog.domain.Comment;
 
 import java.util.List;
